+++ tests/integration/test_media_player_vlc.py
@@ -38,6 +38,9 @@ class MediaPlayerVlcIntegrationTestCase(TestCasePoller):
             "--text-renderer=tdummy",
         ]
 
+        # use default window
+        self.use_default_window = True
+
         # create fullscreen flag
         self.fullscreen = True
 
@@ -109,6 +112,7 @@ class MediaPlayerVlcIntegrationTestCase(TestCasePoller):
                 "vlc": {
                     "instance_parameters": self.instance_parameters,
                     "media_parameters": self.media_parameters,
+                    "use_default_window": self.use_default_window,
                 },
             }